Node type selection while FQ creating
