+++ tools/pot/openvino/tools/pot/algorithms/quantization/utils.py
@@ -245,33 +245,6 @@ def get_quantize_op_config(op, config, opt_conf=None):
     return qconfig
 
 
-def get_hardware_config_operation_type(node, available_types):
-    """ This function gets type by child
-    for hardware configuration of FQ node
-    :param node: node-type object
-    :param available_types: available types with config
-    :return: default or special type of layer as string
-    """
-
-    def _is_depth_wise(node):
-        if node.type == 'Convolution' and node.has_valid('group'):
-            group = node['group']
-            output = node['output']
-            input_shape = get_input_shape(node, 0)
-            if group == output and input_shape[1] == output:
-                return True
-        return False
-
-    type_checkers = {
-        'DepthWiseConvolution': _is_depth_wise
-    }
-
-    for real_type in type_checkers:
-        if real_type in available_types and type_checkers[real_type](node):
-            return real_type
-    return node.type
-
-
 def get_tensor_statistics(range_estimator_config, for_weights, **kwargs):
     stats = {}
     for stats_name in ['min', 'max']:
